Styles cards with yellow border, icons etc
Also changed the height mechanic to html flexbox so that the body can
expand if more cards are added, but is a min of 100vh. And on this note I
put the announcements and trending into a single box so that trending
doesn't get dragged down when more cards are added.